requirements.txt
streamlit-1.36.0-py2.py3-none-any.whl
transformers
torch
numpy==1.24.1
openai
plotly